requirements.txt
requests==2.13.0
tabulate==0.7.7
python-dateutil==2.6.0
